Refactoring: sidebar
